Добавлена TotalBar. Требует доработки бизнес логики и рассчетов. Добавлен диалог создания клиентов. Сделаны проверки на присутствие клиента в Order при удалении. Не удаляется, если присутствуют документ ыпо данному клиенту.

+++ Native/src/components/screens/ClientScreen/ClientScreen.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from 'react'
-import { Alert, FlatList, Modal, ScrollView, StyleSheet, View } from 'react-native'
+import { Alert, FlatList, StyleSheet, View } from 'react-native'
 import axios from 'axios'
 import { IClient } from '../../../../../Interfaces/IClient'
 import { ClientCard } from './ClientCard/ClientCard'
@@ -7,15 +7,23 @@ import { ClientForm } from './ClientForm/ClientForm'
 import { TextField } from '../../../shared/TextField'
 import { SETTINGS } from '../../../Default'
 import { useContextData } from '../../../ContextProvider'
+import { BottomBar } from '../../BottomBar'
+
+interface IClientScreenProps {
+  select?: (clinet: IClient) => void
+}
 
 
-export const ClientScreen = () => {
+
+export const ClientScreen = ({ select }: IClientScreenProps) => {
 
 
   const { clients, setClients } = useContextData()
-  
+
   const [selectClient, setSelectClient] = useState<IClient>()
   const [openModal, setOpenModal] = useState(false)
+
+  //Фильтрация списка
   const [filterClient, setFilterClient] = useState<IClient[]>([])
   const [searchText, setSearchText] = useState('')
 
@@ -26,11 +34,28 @@ export const ClientScreen = () => {
   const handleSaveClient = async () => {
     if (selectClient) {
       try {
-        await axios.put(`${server}/api/client/${selectClient._id}`, selectClient)
-        setClients((prevClients) => prevClients.map((client) =>
-          client._id === selectClient._id ? selectClient : client
-        )
-        )
+
+        if (!selectClient.name) {
+          Alert.alert('Имя клиента не может быть пустым')
+          return
+        }
+
+        if (selectClient._id) {
+          await axios.put(`${server}/api/client/${selectClient._id}`, selectClient)
+          setClients((prevClients) =>
+            prevClients.map((client) =>
+              client._id === selectClient._id ? selectClient : client
+            )
+          )
+        } else {
+
+          const response = await axios.post(`${server}/api/client/`, selectClient)
+          const updateClient = response.data
+
+          const updateClients = { ...clients, ...updateClient }
+
+          setClients(updateClients)
+        }
         closeClientForm()
       } catch (error) {
         Alert.alert("Ошибка сохранения данных");
@@ -39,18 +64,44 @@ export const ClientScreen = () => {
     }
   }
 
-  const handleDeleteClient = async () => {
-    if (selectClient) {
-      try {
-        await axios.delete(`${server}/api/client/${selectClient._id}`);
-        setClients((prevClients) =>
-          prevClients.filter((client) => client._id !== selectClient._id)
-        );
-        closeClientForm();
-      } catch (error) {
-        Alert.alert("Ошибка удаления клиента:");
-      }
+  const handleOpenNewClientForm = () => {
+    const newClient: IClient = {
+      address: '',
+      gender: '',
+      gps: '',
+      name: '',
+      percent: 15,
+      phone: ''
     }
+    setSelectClient(newClient)
+    setOpenModal(true)
+  }
+
+  const handleDeleteClient = async () => {
+    Alert.alert('Предупреждение', 'Вы действительно хотите удалить этого клиента?', [
+      {
+        text: 'Удалить', onPress: async () => {
+          if (selectClient) {
+            try {
+              const response = await axios.get(`${server}/api/order/getOrderByClientId/${selectClient._id}`)
+              if (response.data.length > 0) {
+                Alert.alert('Не возможно удалить клиента!!!', `Данный клиент используется в ${response.data.length}  операциях`)
+                return
+              }
+              await axios.delete(`${server}/api/client/${selectClient._id}`);
+              setClients((prevClients) =>
+                prevClients.filter((client) => client._id !== selectClient._id)
+              );
+              closeClientForm();
+            } catch (error) {
+              Alert.alert("Ошибка удаления клиента:");
+            }
+          }
+        }
+      },
+      { text: 'Отмена' }
+
+    ],)
   };
 
   const selectClientHandler = (client: IClient) => {
@@ -87,7 +138,7 @@ export const ClientScreen = () => {
     responsDB()
     setFilterClient(clients)
 
-  }, [])
+  }, [clients])
 
   return (
     <View style={style.content}>
@@ -109,10 +160,14 @@ export const ClientScreen = () => {
                   .toLocaleLowerCase()
                 )))
           }
-          renderItem={({ item }) => <ClientCard client={item} select={selectClientHandler} />}
+          renderItem={({ item }) => (!select ?
+            <ClientCard client={item} select={selectClientHandler} /> :
+            <ClientCard client={item} select={select} />
+          )}
           keyExtractor={(item, index) => item._id ?? index.toString()}
         />
       ) : null}
+      <BottomBar bottomData={[{ icons: 'add', action: () => handleOpenNewClientForm() }]} />
       {selectClient ?
         <ClientForm
           openModal={openModal}
